# apps/logistics-backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.ErrorResponse:
    properties:
      detail:
        description: optional internal error
        example: validation failed on field 'email'
        type: string
      error:
        description: user-friendly message
        example: Invalid request
        type: string
    type: object
  inventory.CreateInventoryRequest:
    properties:
      admin_id:
        description: Foreign key
        type: string
      category:
        description: e.g. “Dairy”
        type: string
      description:
        type: string
      images:
        description: could be JSON array or URLs
        type: string
      location:
        description: optional
        type: string
      name:
        description: e.g. “Fresh Milk”
        type: string
      packaging:
        description: “Bucket/Single”
        type: string
      price:
        type: number
      slug:
        type: string
      stock:
        type: integer
      unit:
        description: '"per litre", "per bucket"'
        type: string
    required:
    - category
    - description
    - images
    - location
    - name
    - packaging
    - price
    - slug
    - stock
    - unit
    type: object
  inventory.Inventory:
    properties:
      admin_id:
        description: Foreign key
        type: string
      category:
        description: e.g. “Dairy”
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      images:
        description: could be JSON array or URLs
        type: string
      location:
        description: optional
        type: string
      name:
        description: e.g. “Fresh Milk”
        type: string
      packaging:
        description: “Bucket/Single”
        type: string
      price:
        type: number
      slug:
        description: product slug
        type: string
      stock:
        type: integer
      unit:
        description: '"per litre", "per bucket"'
        type: string
      updated_at:
        type: string
    type: object
  inventory.InventorySummary:
    properties:
      image:
        type: string
      in_stock:
        type: integer
      name:
        type: string
      packaging:
        type: string
      price:
        type: number
      unit:
        type: string
    type: object
  inventory.StorePublicView:
    properties:
      admin_name:
        type: string
      category:
        type: string
      location:
        type: string
      products:
        items:
          $ref: '#/definitions/inventory.InventorySummary'
        type: array
    type: object
  notification.CreateNotificationRequest:
    properties:
      message:
        type: string
      type:
        $ref: '#/definitions/notification.NotificationType'
      user_id:
        type: string
    type: object
  notification.Notification:
    properties:
      id:
        type: string
      message:
        type: string
      sent_at:
        type: string
      type:
        $ref: '#/definitions/notification.NotificationType'
      user_id:
        type: string
    type: object
  notification.NotificationType:
    enum:
    - email
    - sms
    - push
    type: string
    x-enum-varnames:
    - Email
    - SMS
    - Push
...
host: 192.168.100.18:8000
info:
  contact: {}
  description: This is the API for logistics operations.
  title: Logistics API
  version: "1.0"
paths:
...
  /inventories/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently deletes an inventory by their ID
      parameters:
      - description: Inventory ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Inventory deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid inventory ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - JWT: []
      summary: Delete an inventory
      tags:
      - inventories
  /inventories/all_inventories:
    get:
      description: List all inventories with optional pagination
      parameters:
      - description: Limit number of items
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inventory.Inventory'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - JWT: []
      summary: List all inventories
      tags:
      - inventories
  /inventories/by-category:
    get:
      description: Get all inventory items in a specific category
      parameters:
      - description: Category Name
        in: query
        name: category
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inventory.Inventory'
            type: array
        "400":
          description: Invalid Category
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - JWT: []
      summary: Get inventories by category
      tags:
      - inventories
  /inventories/by-id:
    get:
      description: Get a specific inventory item by UUID
      parameters:
      - description: Inventory ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inventory.Inventory'
        "400":
          description: Invalid inventory ID or request body
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - JWT: []
      summary: Get inventory by ID
      tags:
      - inventories
  /inventories/by-name:
    get:
      description: Search inventory by item name (exact match)
      parameters:
      - description: Inventory Name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inventory.Inventory'
        "400":
          description: Invalid inventory ID or request body
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - JWT: []
      summary: Get inventory by name
      tags:
      - inventories
  /inventories/categories:
    get:
      description: Get all unique categories that belong to existing inventories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - JWT: []
      summary: List all categories
      tags:
      - inventories
  /inventories/create:
    post:
      consumes:
      - application/json
      description: Creates an inventory and returns the created object
      parameters:
      - description: Inventory input
        in: body
        name: inventory
        required: true
        schema:
          $ref: '#/definitions/inventory.CreateInventoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/inventory.Inventory'
        "400":
          description: Invalid inventory ID or request body
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - JWT: []
      summary: Create a new inventory
      tags:
      - inventories
  /notifications/{id}:
    get:
      description: Retrieve a notification by their ID
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notification.Notification'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Notification not found
          schema:
            type: string
      security:
      - JWT: []
      summary: Get notification by ID
      tags:
      - notifications
  /notifications/all_notifications:
    get:
      description: Get a list of all notifications
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/notification.Notification'
            type: array
      security:
      - JWT: []
      summary: List all notifications
      tags:
      - notifications
  /notifications/create:
    post:
      consumes:
      - application/json
      description: Create a new notification with user_id, message, etc.
      parameters:
      - description: User Input
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/notification.CreateNotificationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/notification.Notification'
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Failed to create notification
          schema:
            type: string
      security:
      - JWT: []
      summary: Create a new notification
      tags:
      - notifications
...
  /public/store/{adminSlug}:
    get:
      description: View all public products for an admin store by slug
      parameters:
      - description: Admin Slug
        in: path
        name: adminSlug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inventory.StorePublicView'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: View public store page
      tags:
      - public
  /public/store/{adminSlug}/product/{productSlug}:
    get:
      description: View a specific product by admin slug and product slug
      parameters:
      - description: Admin Slug
        in: path
        name: adminSlug
        required: true
        type: string
      - description: Product Slug
        in: path
        name: productSlug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inventory.Inventory'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: View public product page
      tags:
      - public
...
schemes:
- http
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"